Move BoardSerialization into board package
