+++ src/app/backend/board/board-serialization.spec.ts
@@ -1,7 +1,7 @@
-import {Board} from './board/board';
-import {Direction, Enemy, Mover} from './cells';
+import {Board} from './board';
+import {Direction, Enemy, Mover} from '../cells';
 import {BoardSerialization} from './board-serialization';
-import {Coordinate} from './board/coordinate';
+import {Coordinate} from './coordinate';
 
 describe('backend/board-serialization', () => {
   it('serialize() works.', () => {
